refactor: Rename `lib/injectLogsIamRole.js`
