+++ lib/inject-logs-iam-role.js
