+++ lib/inject-logs-iam-role.test.js
@@ -20,7 +20,7 @@ describe('injectLogsIamRole', () => {
       },
     };
 
-    injectLogsIamRole = proxyquire('./injectLogsIamRole', {
+    injectLogsIamRole = proxyquire('./inject-logs-iam-role', {
       '@serverless/platform-client': platformClientStub,
     });
   });
