+++ lib/plugin.js
@@ -8,7 +8,7 @@ const logsCollection = require('./logsCollection');
 const login = require('./login');
 const logout = require('./logout');
 const wrap = require('./wrap');
-const injectLogsIamRole = require('./injectLogsIamRole');
+const injectLogsIamRole = require('./inject-logs-iam-role');
 const injectOutputOutputs = require('./injectOutputOutputs');
 const wrapClean = require('./wrapClean');
 const getAppUids = require('./app-uids');
