+++ lib/plugin.test.js
@@ -50,7 +50,7 @@ describe('plugin', () => {
       './deployment': { saveDeployment, createAndSetDeploymentUid },
       './generate-event': { eventDict: {}, generate },
       './deploy-profile': { configureDeployProfile },
-      './injectLogsIamRole': injectLogsIamRole,
+      './inject-logs-iam-role': injectLogsIamRole,
       './isAuthenticated': () => true,
       './setApiGatewayAccessLogFormat': setApiGatewayAccessLogFormat,
       '@serverless/config/utils': {
